requirements.txt
torch
transformers>=4.42.0
datasets
peft
accelerate
bitsandbytes
pillow
sentencepiece
timm>=0.9.16
torchvision
